check that other users can't edit wrong info

+++ api/src/main/java/fhnw/dreamteam/stockstracker/service/CurrencyService.java
@@ -1,6 +1,7 @@
 package fhnw.dreamteam.stockstracker.service;
 
 import fhnw.dreamteam.stockstracker.data.models.Currency;
+import fhnw.dreamteam.stockstracker.data.models.Stock;
 import fhnw.dreamteam.stockstracker.data.repository.CurrencyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -43,6 +44,11 @@ public class CurrencyService {
      */
     public Currency editCurrency(@Valid final Currency currency) throws Exception {
         Optional<Currency> dbCurrency = currencyRepository.findById(currency.getId());
+
+        if(dbCurrency.isPresent() && userService.getCurrentUser().getUsername() != dbCurrency.get().getUser().getUsername()) {
+            throw new Exception("You are not permitted to edit this currency");
+        }
+
         if (currency.getId() != null && dbCurrency != null && dbCurrency.isPresent()) {
             dbCurrency.get().setName(currency.getName());
             return currencyRepository.save(dbCurrency.get());
@@ -56,26 +62,15 @@ public class CurrencyService {
      *
      * @param currencyID The stock's ID to delete.
      */
-    public void deleteCurrency(final Long currencyID) {
+    public void deleteCurrency(final Long currencyID) throws Exception {
+        Optional<Currency> dbCurrency = currencyRepository.findById(currencyID);
+
+        if(dbCurrency.isPresent() && userService.getCurrentUser().getUsername() != dbCurrency.get().getUser().getUsername()) {
+            throw new Exception("You are not permitted to delete this currency");
+        }
+
         currencyRepository.deleteById(currencyID);
     }
-
-    // /**
-    //  * Find a stock by it's ID.
-    //  *
-    //  * @param stockId The {@link Stock}'s ID.
-    //  *
-    //  * @return Returns the stock.
-    //  *
-    //  * @throws Exception Throws if nothing found.
-    //  */
-    // public Stock findStockById(final Long stockId) throws Exception {
-    //     List<Stock> stockList = stockRepository.findById(stockId);
-    //     if (stockList.isEmpty()) {
-    //         throw new Exception("No customer with ID " + stockId + " found.");
-    //     }
-    //     return stockList.get(0);
-    // }
 
     /**
      * Get all currencies.
